refactor: dynamic properties method moved to PluginService
