+++ rundeckapp/grails-app/services/rundeck/services/FrameworkService.groovy
@@ -814,87 +814,6 @@ class FrameworkService implements ApplicationContextAware, AuthContextProcessor,
     }
 
     /**
-     * load the dynamic select values for properties from the plugin, or null
-     * @param serviceName
-     * @param type
-     * @param project
-     * @param services
-     * @return
-     */
-    def Map<String, Object> getDynamicProperties(
-        String serviceName,
-        String type,
-        String project,
-        Services services
-    ) {
-        final PropertyResolver resolver = PropertyResolverFactory.createPluginRuntimeResolver(
-            project,
-            rundeckFramework,
-            null,
-            serviceName,
-            type
-        );
-
-        def pluginServiceType
-
-
-        if(serviceName == ServiceNameConstants.WorkflowNodeStep){
-            pluginServiceType = rundeckFramework.getNodeStepExecutorService()
-        }else if(serviceName == ServiceNameConstants.WorkflowStep){
-            pluginServiceType = rundeckFramework.getStepExecutionService()
-        }else{
-            pluginServiceType = serviceName
-        }
-
-        def pluginDescriptor = pluginService.getPluginDescriptor(type, pluginServiceType)
-
-        final Map<String, Object> config = PluginAdapterUtility.mapDescribedProperties(
-            resolver,
-            pluginDescriptor.description,
-            PropertyScope.Project
-        )
-
-        Description desc = pluginDescriptor.description
-
-        //add custom mapping for plugin properties at framework level.
-        // eg: set the value of the plugin properties based on framework setting
-        //     builder.mapping("framework.path.attr", "attr")
-        Map frameworkProperties = rundeckFramework.getPropertyLookup().getPropertiesMap()
-        if(desc.getFwkPropertiesMapping()){
-            Map props = Validator.performMapping(frameworkProperties, desc.getFwkPropertiesMapping(), true)
-
-            if(props){
-                props.each {key, value->
-                    if(!config.get(key)){
-                        config.put(key, value)
-                    }
-                }
-            }
-        }
-
-        //add custom mapping for plugin properties at project level.
-        // eg: set the value of the plugin properties based on project setting
-        //     builder.mapping("project.path.attr", "attr")
-        Map projectProperties = getFrameworkProject(project).getProperties()
-        if(desc.getPropertiesMapping()){
-            Map props = Validator.performMapping(projectProperties, desc.getPropertiesMapping(), true)
-
-            if(props){
-                props.each {key, value->
-                    if(!config.get(key)){
-                        config.put(key, value)
-                    }
-                }
-            }
-        }
-
-        def plugin = pluginDescriptor.instance
-        if(plugin instanceof DynamicProperties){
-            return plugin.dynamicProperties(config, services)
-        }
-        return null
-    }
-    /**
      * Return the list of NodeStepPlugin descriptions
      * @param framework
      * @return
